Fixed ReSharper violations.

+++ SharpSnmpLib/Mib/Lexer.cs
@@ -177,7 +177,7 @@ namespace Lextm.SharpSnmpLib.Mib
                 case '\r':
                     return false;
                 default:
-                    if ((int)current == 0x1A)
+                    if (current == 0x1A)
                     {
                         // IMPORTANT: ignore invisible characters such as SUB.
                         return false;
